pyproject.toml
[project]
name = "zelda-miloutte"
version = "0.1.0"
description = "A simple Zelda-like 2D top-down adventure game starring Miloutte"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
zelda-miloutte = "zelda_miloutte.main:main"
